requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

# Image Processing
opencv-python
numpy
Pillow

# Machine Learning
tensorflow
keras

# OCR
pytesseract
pandas
textdistance
